Pull github auth functionality into a separate module
